FormattableContent: Extracting formatting code.
Creating FormattableContentFormatter class to get and apply format styles
